Move default CAFTOP values out of reducer in prep for globalState change
