+++ src/Steps/Distribution/Fields/DSO.Validation.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { Distribution } from "stateManagement/reducer";
+import { Distribution } from "api/CAFTOP/defaults";
 import { populateWithDefaultValue } from "utilities/Validation";
 
 /** Rule for when the DSO is Not Applicable */
