+++ src/Steps/LRDP/Fields/LRDP.Validation.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 import { CAFTOPLRDP } from "api/CAFTOP/types";
-import { LRDP as LRDPDefault } from "stateManagement/reducer";
+import { LRDP as LRDPDefault } from "api/CAFTOP/defaults";
 import { populateWithDefaultValue } from "utilities/Validation";
 
 const lrdpName = z.string().trim().max(50, "Name cannot exceed 50 characters");
